Fix isModified behavior of Configurable
Only when isModified sometimes returns false, the IDE recognizes changes
as "applied" and persists the configuration.

We compare the TimestampFormat value based on their title only: getting
a TimestampFormat from the TimestampFormatMap always returns a new
instance, so the comparison on objects would not work.
Other ways to compare the values are possible and might be required in
the future.